Rename formatter/ to fmt/ for brevity
