+++ src/plugins/fmt/index.js
